requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
psutil>=5.8.0
requests>=2.25.0
websockets>=10.0
pika>=1.2.0
psycopg2-binary>=2.9.0

